% Auto-generated: do not edit by hand
\name{'DashNgl'DashNgl}

\alias{'DashNgl'DashNgl}

\title{DashNgl component}

\description{
Dash ngl component
}

\usage{
'DashNgl'DashNgl(id=NULL, viewportStyle=NULL, stageParameters=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{viewportStyle}{Named list. CSS styling for viewport container}

\item{stageParameters}{Named list. Parameters for the stage}

\item{data}{List of lists containing elements 'selectedvalue', 'chain', 'color', 'filename', 'ext', 'config'.
those elements have the following types:
  - selectedvalue (character; required)
  - chain (character; required)
  - color (character; required)
  - filename (character; required)
  - ext (character; optional)
  - config (optional): . config has the following type: lists containing elements 'type', 'input'.
those elements have the following types:
  - type (character; required)
  - input (unnamed list | named list | character; optional)s | lists containing elements 'selectedvalue', 'chain', 'color', 'filename', 'ext', 'config'.
those elements have the following types:
  - selectedvalue (character; required)
  - chain (character; required)
  - color (character; required)
  - filename (character; required)
  - ext (character; optional)
  - config (optional): . config has the following type: lists containing elements 'type', 'input'.
those elements have the following types:
  - type (character; required)
  - input (unnamed list | named list | character; optional). Custom property}
}
